finished all the works related to category and tags

+++ app/Http/Controllers/CategoriesController.php
@@ -97,6 +97,12 @@ class CategoriesController extends Controller
      */
     public function destroy(Category $category)
     {
+        if($category->posts->count()>0)
+        {
+            session()->flash('error','Category cannot be deleted because it has some posts');
+
+            return redirect()->back();
+        }
         $category->delete();
 
         session()->flash('success','category deleted successfully');
