+++ app/Http/Controllers/TagsController.php
@@ -97,6 +97,10 @@ class TagsController extends Controller
      */
     public function destroy(Tag $tag)
     {
+        if($tag->posts->count()>0){
+            session()->flash('error','Tag cannot be deleted because it has some posts');
+            return redirect()->back();
+        }
         $tag->delete();
 
         session()->flash('success','tag deleted successfully');
